requirements.txt
langchain
python-dotenv
ipykernel 
langchain_community
pypdf
bs4
arxiv
wikipedia
langchain-text-splitters
langchain-openai
sentence_transformers
langchain_huggingface
faiss-cpu
duckdb
openai
langchain-groq
duckduckgo_search
pymupdf
validators
youtube_transcript_api
unstructured
pytube
huggingface_hub
streamlit
Flask
Flask-Bcrypt
Flask-SQLAlchemy
pytesseract
statsmodels
huggingface_hub
ollama
PyPDF2
requests
psycopg2
openai
sentence-transformers